Add files via upload
Added new classes and updated some of the code to load the maze from the ScreenCustomizer class. Still working on implementing the maze into the SC class. Also, has a nasty error when the maze close button is pressed. However, believe that will resolve itself after it is switched from the JPane to the javafx scene.